pyproject.toml
[tool.poetry]
name = "liberandum-fastapi-frontend"
version = "0.1.0"
description = ""
authors = ["Александр <95116155+GaevoyAlex@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = ">=0.115.12,<0.116.0"
sqlalchemy = ">=2.0.40,<3.0.0"
python-jose = ">=3.4.0,<4.0.0"
passlib = ">=1.7.4,<2.0.0"
httpx = ">=0.28.1,<0.29.0"
uvicorn = ">=0.34.2,<0.35.0"
alembic = ">=1.15.2,<2.0.0"
python-multipart = "^0.0.7"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.6.3"}
pydantic-settings = "^2.2.1"  # Добавьте эту строку
python-dotenv = "^1.0.1"
google = "^3.0.0"
google-auth = "^2.40.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

